# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Excel Add-in API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ExecuteCommandBody {
  /** The command to execute (e.g., "Sum", "Multiply", "Average", "Min", "Max", "Subtract", "Divide") */
  command: string;
  /** The numeric values to operate on */
  values: number[];
  /**
   * Optional cell range reference (e.g., "A1:C3")
   * @nullable
   */
  cellRange?: string | null;
}

export interface CommandResult {
  command: string;
  result: number;
  /** The formula string representation */
  formula: string;
  description: string;
}

export interface CommandHistoryItem {
  id: number;
  command: string;
  values: number[];
  result: number;
  formula: string;
  /** @nullable */
  cellRange: string | null;
  executedAt: string;
}

export interface CellDimensionRequest {
  /** The text content of the cell */
  text: string;
  /**
   * Font size in points (default 11)
   * @nullable
   */
  fontSize?: number | null;
  /**
   * Whether the text is bold
   * @nullable
   */
  bold?: boolean | null;
  /**
   * Cell padding in pixels (default 8)
   * @nullable
   */
  padding?: number | null;
  /**
   * Maximum cell width in pixels
   * @nullable
   */
  maxWidth?: number | null;
  /**
   * Whether to wrap text (affects height calculation)
   * @nullable
   */
  wrapText?: boolean | null;
}

export interface CellDimensions {
  /** Calculated cell width in pixels */
  width: number;
  /** Calculated cell height in pixels */
  height: number;
  /** Width in Excel column units */
  widthExcel: number;
  /** Height in Excel row units (points) */
  heightExcel: number;
  characterCount: number;
  lineCount: number;
}

export interface BatchCellDimensionRequest {
  cells: CellDimensionRequest[];
  /**
   * If true, returns the max width for all cells
   * @nullable
   */
  uniformWidth?: boolean | null;
  /**
   * If true, returns the max height for all cells
   * @nullable
   */
  uniformHeight?: boolean | null;
}

export interface BatchCellDimensions {
  results: CellDimensions[];
  maxWidth: number;
  maxHeight: number;
  recommendedWidth: number;
  recommendedHeight: number;
}

export interface ErrorResponse {
  error: string;
}