# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Excel Add-in API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Interprets and executes commands like Sum, Multiply, Average, Min, Max on provided values
 * @summary Execute a spreadsheet command
 */
export const ExecuteCommandBody = zod.object({
  command: zod
    .string()
    .describe(
      'The command to execute (e.g., \"Sum\", \"Multiply\", \"Average\", \"Min\", \"Max\", \"Subtract\", \"Divide\")',
    ),
  values: zod.array(zod.number()).describe("The numeric values to operate on"),
  cellRange: zod
    .string()
    .nullish()
    .describe('Optional cell range reference (e.g., \"A1:C3\")'),
});

export const ExecuteCommandResponse = zod.object({
  command: zod.string(),
  result: zod.number(),
  formula: zod.string().describe("The formula string representation"),
  description: zod.string(),
});

/**
 * Returns the last 20 executed commands with their results
 * @summary Get command history
 */
export const GetCommandHistoryResponseItem = zod.object({
  id: zod.number(),
  command: zod.string(),
  values: zod.array(zod.number()),
  result: zod.number(),
  formula: zod.string(),
  cellRange: zod.string().nullable(),
  executedAt: zod.coerce.date(),
});
export const GetCommandHistoryResponse = zod.array(
  GetCommandHistoryResponseItem,
);

/**
 * Calculates the optimal width and height for a cell based on text content and font settings
 * @summary Calculate cell dimensions
 */
export const CalculateCellDimensionsBody = zod.object({
  text: zod.string().describe("The text content of the cell"),
  fontSize: zod.number().nullish().describe("Font size in points (default 11)"),
  bold: zod.boolean().nullish().describe("Whether the text is bold"),
  padding: zod
    .number()
    .nullish()
    .describe("Cell padding in pixels (default 8)"),
  maxWidth: zod.number().nullish().describe("Maximum cell width in pixels"),
  wrapText: zod
    .boolean()
    .nullish()
    .describe("Whether to wrap text (affects height calculation)"),
});

export const CalculateCellDimensionsResponse = zod.object({
  width: zod.number().describe("Calculated cell width in pixels"),
  height: zod.number().describe("Calculated cell height in pixels"),
  widthExcel: zod.number().describe("Width in Excel column units"),
  heightExcel: zod.number().describe("Height in Excel row units (points)"),
  characterCount: zod.number(),
  lineCount: zod.number(),
});

/**
 * Calculates optimal dimensions for a batch of cells based on their text content
 * @summary Calculate dimensions for multiple cells
 */
export const CalculateBatchCellDimensionsBody = zod.object({
  cells: zod.array(
    zod.object({
      text: zod.string().describe("The text content of the cell"),
      fontSize: zod
        .number()
        .nullish()
        .describe("Font size in points (default 11)"),
      bold: zod.boolean().nullish().describe("Whether the text is bold"),
      padding: zod
        .number()
        .nullish()
        .describe("Cell padding in pixels (default 8)"),
      maxWidth: zod.number().nullish().describe("Maximum cell width in pixels"),
      wrapText: zod
        .boolean()
        .nullish()
        .describe("Whether to wrap text (affects height calculation)"),
    }),
  ),
  uniformWidth: zod
    .boolean()
    .nullish()
    .describe("If true, returns the max width for all cells"),
  uniformHeight: zod
    .boolean()
    .nullish()
    .describe("If true, returns the max height for all cells"),
});

export const CalculateBatchCellDimensionsResponse = zod.object({
  results: zod.array(
    zod.object({
      width: zod.number().describe("Calculated cell width in pixels"),
      height: zod.number().describe("Calculated cell height in pixels"),
      widthExcel: zod.number().describe("Width in Excel column units"),
      heightExcel: zod.number().describe("Height in Excel row units (points)"),
      characterCount: zod.number(),
      lineCount: zod.number(),
    }),
  ),
  maxWidth: zod.number(),
  maxHeight: zod.number(),
  recommendedWidth: zod.number(),
  recommendedHeight: zod.number(),
});