# lib/api-zod/src/generated/types/batchCellDimensionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Excel Add-in API
 * OpenAPI spec version: 0.1.0
 */
import type { CellDimensionRequest } from "./cellDimensionRequest";

export interface BatchCellDimensionRequest {
  cells: CellDimensionRequest[];
  /**
   * If true, returns the max width for all cells
   * @nullable
   */
  uniformWidth?: boolean | null;
  /**
   * If true, returns the max height for all cells
   * @nullable
   */
  uniformHeight?: boolean | null;
}