# lib/api-zod/src/generated/types/batchCellDimensions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Excel Add-in API
 * OpenAPI spec version: 0.1.0
 */
import type { CellDimensions } from "./cellDimensions";

export interface BatchCellDimensions {
  results: CellDimensions[];
  maxWidth: number;
  maxHeight: number;
  recommendedWidth: number;
  recommendedHeight: number;
}