# lib/api-zod/src/generated/types/cellDimensionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Excel Add-in API
 * OpenAPI spec version: 0.1.0
 */

export interface CellDimensionRequest {
  /** The text content of the cell */
  text: string;
  /**
   * Font size in points (default 11)
   * @nullable
   */
  fontSize?: number | null;
  /**
   * Whether the text is bold
   * @nullable
   */
  bold?: boolean | null;
  /**
   * Cell padding in pixels (default 8)
   * @nullable
   */
  padding?: number | null;
  /**
   * Maximum cell width in pixels
   * @nullable
   */
  maxWidth?: number | null;
  /**
   * Whether to wrap text (affects height calculation)
   * @nullable
   */
  wrapText?: boolean | null;
}