# lib/api-zod/src/generated/types/cellDimensions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Excel Add-in API
 * OpenAPI spec version: 0.1.0
 */

export interface CellDimensions {
  /** Calculated cell width in pixels */
  width: number;
  /** Calculated cell height in pixels */
  height: number;
  /** Width in Excel column units */
  widthExcel: number;
  /** Height in Excel row units (points) */
  heightExcel: number;
  characterCount: number;
  lineCount: number;
}