# lib/api-zod/src/generated/types/commandHistoryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Excel Add-in API
 * OpenAPI spec version: 0.1.0
 */

export interface CommandHistoryItem {
  id: number;
  command: string;
  values: number[];
  result: number;
  formula: string;
  /** @nullable */
  cellRange: string | null;
  executedAt: Date;
}