# lib/api-zod/src/generated/types/commandResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Excel Add-in API
 * OpenAPI spec version: 0.1.0
 */

export interface CommandResult {
  command: string;
  result: number;
  /** The formula string representation */
  formula: string;
  description: string;
}