# lib/api-zod/src/generated/types/executeCommandBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Excel Add-in API
 * OpenAPI spec version: 0.1.0
 */

export interface ExecuteCommandBody {
  /** The command to execute (e.g., "Sum", "Multiply", "Average", "Min", "Max", "Subtract", "Divide") */
  command: string;
  /** The numeric values to operate on */
  values: number[];
  /**
   * Optional cell range reference (e.g., "A1:C3")
   * @nullable
   */
  cellRange?: string | null;
}